Refactored `ForwardDDBot` to new files Fixes #13
